% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abc}
\alias{abc}
\title{ABC function}
\usage{
abc(observedData, numParticles, numIters, epsilon = 0.1)
}
\arguments{
\item{observedData}{Observed data (e.g., as an Armadillo matrix or vector)}

\item{numParticles}{Number of particles/samples to generate}

\item{numIters}{Number of iterations}

\item{epsilon}{Tolerance threshold}
}
\value{
posteriorSamples Generated posterior parameter samples (e.g., as an Armadillo matrix)
}
\description{
ABC function
}
