% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abcRej}
\alias{abcRej}
\title{ABC function}
\usage{
abcRej(observedData, numParticles, epsilon, priorMin, priorMax)
}
\arguments{
\item{observedData}{Observed data (e.g., as an Armadillo matrix or vector)}

\item{numParticles}{Number of particles/samples to generate}

\item{epsilon}{Tolerance threshold}

\item{priorMin}{Lower bound of the prior distribution}

\item{priorMax}{Upper bound of the prior distribution}
}
\value{
posteriorSamples Generated posterior parameter samples (e.g., as an Armadillo matrix)
}
\description{
ABC function
}
