% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Function to calculate the euclidean distance between observed and simulated data}
\usage{
calculateDistance(observedData, simulatedData)
}
\arguments{
\item{observedData}{Matrix of observed data (each row represents a time point)}

\item{simulatedData}{Cube of simulated data (each second dimension represents a simulated data point)}
}
\value{
distances Vector of distances between observed and simulated data points
}
\description{
Function to calculate the euclidean distance between observed and simulated data
}
