% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateSimulatedData}
\alias{generateSimulatedData}
\title{Function to generate simulated data from the zombie outbreak model}
\usage{
generateSimulatedData(parameters, numTimePoints, starting)
}
\arguments{
\item{parameters}{A matrix of parameter values (each row represents a set of parameter values)}

\item{numTimePoints}{Number of time points to simulate}

\item{starting}{State of the population in first time point, 0 - Susceptible, 1 - Zombie - 2, 3 - Removed}
}
\value{
A 3D array of simulated data with dimensions (numParticles, numTimePoints, 3). The third dimension represents the populations: 1 - Susceptible, 2 - Zombie, 3 - Removed
}
\description{
Function to generate simulated data from the zombie outbreak model
}
