% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generateStartCond}
\alias{generateStartCond}
\title{Gives a list of starting conditions of the model}
\usage{
generateStartCond(
  N = NULL,
  initial.inf = NULL,
  beta = 2/(N),
  kappa = 3/(N),
  rho = 1/N
)
}
\arguments{
\item{N}{Size of total population to model}

\item{initial.inf}{Number of infected people in the first time point}

\item{beta}{The bite rate}

\item{kappa}{The kill rate}

\item{rho}{The resurrection rate}
}
\value{
a list with elements
- `populations` A vector with elements showing the counts of the initial susceptible, zombie and removed populations
- `N` The total population size
- `beta` The beta supplied as input, , otherwise defaults to 2/N
- `kappa` The kappa supplied as input, otherwise defaults to 3/N
- `rho` The rho supplied as input, otherwise defaults to 1/N
}
\description{
Gives a list of starting conditions of the model
}
