% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generateSZRdata}
\alias{generateSZRdata}
\title{Gives simulated data for a deterministic SZR model}
\usage{
generateSZRdata(initial_cond, total.T = NULL)
}
\arguments{
\item{initial_cond}{The initial conditions of the simulation, as produced by the `generateStartCond()` functionS}

\item{total.T}{The number of time points to simulate}
}
\value{
A list with elements
- `results` A data frame with `total.T` rows and 4 columns showing the population counts for each time point
- `N` The total population, as input
- `beta` The true bite rate
- `kappa` The true kill rate
- `rho` The true resurrection rate
}
\description{
Gives simulated data for a deterministic SZR model
}
