% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_posterior_samples}
\alias{plot_posterior_samples}
\title{Plots the density for each of the model parameters' ABC accepted samples with a black line indicating the true values}
\usage{
plot_posterior_samples(sample, true_vals)
}
\arguments{
\item{sample}{A data frame with the accepted parameter values as returned by `abcRej`}

\item{true_vals}{The true values of the parameters used to generate the data}
}
\value{
A list with 3 plots for each of the parameters in the SZR model
}
\description{
Plots the density for each of the model parameters' ABC accepted samples with a black line indicating the true values
}
