% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_SZR}
\alias{plot_SZR}
\title{Plot SZR curves}
\usage{
plot_SZR(data)
}
\arguments{
\item{data}{A data frame in the same format as the `results` dataframe returned by `generateSZRdata()`}
}
\value{
A plot showing the changes in the S, Z and R populations over time
}
\description{
Plot SZR curves
}
