\name{ZombieSim-package}
\alias{ZombieSim-package}
\alias{ZombieSim}
\docType{package}
\title{
\packageTitle{ZombieSim}
}
\description{
\packageDescription{ZombieSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ZombieSim}
\packageIndices{ZombieSim}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ZombieSim}

Maintainer: \packageMaintainer{ZombieSim}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\examples{

}
